// app/src/engine/types/Trivia.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Player } from "./Player";
import type { TriviaAnswer } from "./TriviaAnswer";
import type { TriviaPhase } from "./TriviaPhase";
import type { TriviaQuestion } from "./TriviaQuestion";

export type Trivia = { title: string, questions: Array<TriviaQuestion>, 
/**
 * The question asked (or last asked).
 */
current: number, phase: TriviaPhase, askedAt: number, players: Array<Player>, 
/**
 * Answers to the current question.
 */
answers: Array<TriviaAnswer>, joinUrl: string, joinQr: string, showJoin: boolean, };